refactor: migrate TrainingHub progression to module-based instrumentation
Refactored TrainingHub progression tracking to use module-based
instrumentation pattern (matching TransformersTrainer approach) and
added SDK version mismatch warnings.

Changes:
- Created traininghub_progression.py module with progression tracking logic
- Updated traininghub.py wrapper to import from module instead of inline code
- Added SDK version mismatch warning when client and runtime versions differ
- Added version mismatch warning to checkpoint tracking (transformers.py)
- Created comprehensive test suite (9 tests) for traininghub_progression module
- Updated 14 wrapper tests to expect module imports instead of inline code

Benefits:
- Consistent architecture across TransformersTrainer and TrainingHubTrainer
- Better separation of concerns (wrapper generation vs runtime logic)
- Easier to test and maintain progression tracking independently
- Clear warnings when SDK versions don't match

All tests passing: 224 rhai tests, make verify succeeded
